wazuh-ml/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP Requests
requests>=2.31.0
urllib3>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0

# Date/Time utilities
python-dateutil>=2.8.0

# Env management
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0

# LLM (optional)
openai>=1.42.0

# Web framework (optional)
fastapi>=0.100.0
uvicorn>=0.23.0

#torch and metrics
".[torch, metrics]"